Introduce small script improvements

+++ local-k8s.sh
@@ -1,4 +1,4 @@
-#!/bin/bash
+#!/usr/bin/env bash
 set -eou pipefail
 
 # Set the current directory in a variable
@@ -8,11 +8,11 @@ CURRENT_DIR="$(dirname "$0")"
 source "$CURRENT_DIR/utilities/helpers.sh"
 
 REQUIREMENTS=(
-  "kind"
-  "helm"
-  "kubectl"
-  "curl"
-  "nohup"
+    "kind"
+    "helm"
+    "kubectl"
+    "curl"
+    "nohup"
 )
 
 # NOTE: If triggering some of the scripts locally on Mac, you might find an error from the test complaining
@@ -24,6 +24,8 @@ WEAVIATE_PORT=${WEAVIATE_PORT:-8080}
 WEAVIATE_GRPC_PORT=${WEAVIATE_GRPC_PORT:-50051}
 MODULES=${MODULES:-""}
 HELM_BRANCH=${HELM_BRANCH:-""}
+DELETE_STS=${DELETE_STS:-"true"}
+REPLICAS=${REPLICAS:-1}
 PROMETHEUS_PORT=9091
 GRAFANA_PORT=3000
 TARGET=""
@@ -33,12 +35,22 @@ WEAVIATE_IMAGES=(
     "semitechnologies/contextionary:en0.16.0-v1.2.1"
 )
 
-function upgrade_to_raft() {
-    echo "upgrade # Upgrading to RAFT"
+function get_timeout() {
+    # Increase timeout if MODULES is not empty as the module image might take some time to download
+    # and calculate the timeout value based on the number of replicas
+    modules_timeout=0
+    if [ -n "$MODULES" ]; then
+        modules_timeout=1200
+    fi
+    echo "$((modules_timeout + (REPLICAS * 90)))s"
+}
+
+function upgrade() {
+    echo_green "upgrade # Upgrading to Weaviate ${WEAVIATE_VERSION}"
 
     # Upload images to cluster if --local-images flag is passed
     if [ "${1:-}" == "--local-images" ]; then
-        echo "Uploading local images to the cluster"
+        echo_green "Uploading local images to the cluster"
         for image in "${WEAVIATE_IMAGES[@]}"; do
             kind load docker-image $image --name weaviate-k8s
         done
@@ -47,10 +59,16 @@ function upgrade_to_raft() {
     # This function sets up weaviate-helm and sets the global env var $TARGET
     setup_helm $HELM_BRANCH
 
-    echo "upgrade # Deleting Weaviate StatefulSet"
-    kubectl delete sts weaviate -n weaviate
+    if [ "$DELETE_STS" == "true" ]; then
+        echo_yellow "upgrade # Deleting Weaviate StatefulSet"
+        kubectl delete sts weaviate -n weaviate
+    else
+        echo_green "upgrade # Weaviate StatefulSet is not being deleted"
+    fi
 
     HELM_VALUES=$(generate_helm_values)
+    # Configure parallel upgrade, instead of default rolling update
+    # HELM_VALUES="$HELM_VALUES --set updateStrategy.rollingUpdate.maxUnavailable=100%"
 
     VALUES_OVERRIDE=""
     # Check if values-override.yaml file exists
@@ -58,7 +76,7 @@ function upgrade_to_raft() {
         VALUES_OVERRIDE="-f ${CURRENT_DIR}/values-override.yaml"
     fi
 
-    echo -e "upgrade # Upgrading weaviate-helm with values: \n\
+    echo_green "upgrade # Upgrading weaviate-helm with values: \n\
         TARGET: $TARGET \n\
         HELM_VALUES: $(echo "$HELM_VALUES" | tr -s ' ') \n\
         VALUES_OVERRIDE: $VALUES_OVERRIDE"
@@ -66,20 +84,22 @@ function upgrade_to_raft() {
         --namespace weaviate \
         $HELM_VALUES \
         $VALUES_OVERRIDE
-   
+
     # Wait for Weaviate to be up
-    kubectl wait sts/weaviate -n weaviate --for jsonpath='{.status.readyReplicas}'=${REPLICAS} --timeout=100s
+    TIMEOUT=$(get_timeout)
+    echo_green "upgrade # Waiting (with timeout=$TIMEOUT) for Weaviate $REPLICAS node cluster to be ready"
+    kubectl wait sts/weaviate -n weaviate --for jsonpath='{.status.readyReplicas}'=${REPLICAS} --timeout=${TIMEOUT}
     port_forward_to_weaviate
     wait_weaviate
 
     # Check if Weaviate is up
-    curl http://localhost:${WEAVIATE_PORT}/v1/nodes
+    wait_for_all_healthy_nodes $REPLICAS
+    echo_green "upgrade # Success"
 }
 
 
 function setup() {
-
-    echo "setup # Setting up Weaviate on local k8s"
+    echo_green "setup # Setting up Weaviate $WEAVIATE_VERSION on local k8s"
 
     # Create Kind config file
     cat <<EOF > /tmp/kind-config.yaml
@@ -91,13 +111,13 @@ nodes:
 $(for i in $(seq 1 $WORKERS); do echo "- role: worker"; done)
 EOF
 
-    echo "setup # Create local k8s cluster"
+    echo_green "setup # Create local k8s cluster"
     # Create k8s Kind Cluster
     kind create cluster --wait 120s --name weaviate-k8s --config /tmp/kind-config.yaml
 
     # Upload images to cluster if --local-images flag is passed
     if [ "${1:-}" == "--local-images" ]; then
-        echo "Uploading local images to the cluster"
+        echo_green "Uploading local images to the cluster"
         for image in "${WEAVIATE_IMAGES[@]}"; do
             kind load docker-image $image --name weaviate-k8s
         done
@@ -117,7 +137,7 @@ EOF
 
     HELM_VALUES=$(generate_helm_values)
 
-    echo -e "setup # Deploying weaviate-helm with values: \n\
+    echo_green "setup # Deploying weaviate-helm with values: \n\
         TARGET: $TARGET \n\
         HELM_VALUES: $(echo "$HELM_VALUES" | tr -s ' ') \n\
         VALUES_OVERRIDE: $VALUES_OVERRIDE"
@@ -128,29 +148,20 @@ EOF
     $VALUES_OVERRIDE
     #--set debug=true
 
-    # Calculate the timeout value based on the number of replicas
-    if [[ $REPLICAS -le 1 ]]; then
-        TIMEOUT=90s
-    else
-        TIMEOUT=$((REPLICAS * 60))s
-    fi
-
-    # Increase timeout if MODULES is not empty as the module image might take some time to download
-    if [ -n "$MODULES" ]; then
-        TIMEOUT=900s
-    fi
-
     # Wait for Weaviate to be up
+    TIMEOUT=$(get_timeout)
+    echo_green "setup # Waiting (with timeout=$TIMEOUT) for Weaviate $REPLICAS node cluster to be ready"
     kubectl wait sts/weaviate -n weaviate --for jsonpath='{.status.readyReplicas}'=${REPLICAS} --timeout=${TIMEOUT}
     port_forward_to_weaviate
     wait_weaviate
 
     # Check if Weaviate is up
-    curl http://localhost:${WEAVIATE_PORT}/v1/nodes
+    wait_for_all_healthy_nodes $REPLICAS
+    echo_green "setup # Success"
 }
 
 function clean() {
-    echo "clean # Cleaning up local k8s cluster..."
+    echo_green "clean # Cleaning up local k8s cluster..."
 
     # Kill kubectl port-forward processes running in the background
     pkill -f "kubectl-relay" || true
@@ -172,6 +183,7 @@ function clean() {
         # Delete Kind cluster
         kind delete cluster --name weaviate-k8s
     fi
+    echo_green "clean # Success"
 }
 
 
@@ -191,17 +203,18 @@ fi
 
 # Check if all requirements are installed
 for requirement in "${REQUIREMENTS[@]}"; do
-  if ! command -v $requirement &> /dev/null; then
-    echo "Please install '$requirement' before running this script"
-    echo "    brew install $requirement"
-    exit 1
-  fi
+    if ! command -v $requirement &> /dev/null; then
+        echo "Please install '$requirement' before running this script"
+        echo "    brew install $requirement"
+        exit 1
+    fi
 done
 
 # Add an optional second argument --local-images (defaults to false) which allows uploading the local images to the cluster using 
 # kind load docker-image <image-name> --name weaviate-k8s
 LOCAL_IMAGES=""
 if [ $# -ge 2 ] && [ "$2" == "--local-images" ]; then
+    echo "Local images enabled"
     LOCAL_IMAGES="--local-images"
 fi
 
@@ -211,7 +224,7 @@ case $1 in
         setup $LOCAL_IMAGES
         ;;
     "upgrade")
-        upgrade_to_raft $LOCAL_IMAGES
+        upgrade $LOCAL_IMAGES
         ;;
     "clean")
         clean
@@ -225,5 +238,5 @@ esac
 
 # Retrieve Weaviate logs
 if [ $? -ne 0 ]; then
-  kubectl logs -n weaviate -l app.kubernetes.io/name=weaviate
+    kubectl logs -n weaviate -l app.kubernetes.io/name=weaviate
 fi
